pyproject.toml
[project]
name = "quiz-telegram-bot"
version = "1.0.0"
description = "A Telegram bot for English learning through daily quizzes"
authors = [
    {name = "Quiz Bot Team", email = "team@example.com"}
]
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "aiogram>=3.3.0",
    "aiohttp>=3.9.1",
    "google-api-python-client>=2.111.0",
    "google-auth>=2.25.2",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.uv]
package = false
